PIMOB-1349: Refactoring
