+++ frames/src/main/java/com/checkout/frames/component/cardnumber/CardNumberViewModel.kt
@@ -1,14 +1,15 @@
 package com.checkout.frames.component.cardnumber
 
-import androidx.compose.ui.Modifier
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.checkout.base.mapper.Mapper
+import com.checkout.frames.component.base.InputComponentState
 import com.checkout.frames.di.base.InjectionClient
 import com.checkout.frames.di.base.Injector
 import com.checkout.frames.di.component.CardNumberViewModelSubComponent
-import com.checkout.frames.style.component.ContainerStyle
-import com.checkout.frames.style.component.InputComponentStyle
+import com.checkout.frames.style.component.CardNumberComponentStyle
+import com.checkout.frames.style.component.base.InputComponentStyle
+import com.checkout.frames.style.view.InputComponentViewStyle
 import com.checkout.validation.api.CardValidator
 import javax.inject.Inject
 import javax.inject.Provider
@@ -17,13 +18,17 @@ import javax.inject.Provider
 @SuppressWarnings("UnusedPrivateMember")
 internal class CardNumberViewModel @Inject constructor(
     private val cardValidator: CardValidator,
-    private val containerMapper: Mapper<ContainerStyle, Modifier>,
-    private val style: InputComponentStyle
+    private val inputComponentStyleMapper: Mapper<InputComponentStyle, InputComponentViewStyle>,
+    private val inputComponentStateMapper: Mapper<InputComponentStyle, InputComponentState>,
+    private val style: CardNumberComponentStyle
 ) : ViewModel() {
+
+    val viewStyle = inputComponentStyleMapper.map(style.inputStyle)
+    val viewState = inputComponentStateMapper.map(style.inputStyle)
 
     internal class Factory(
         private val injector: Injector,
-        private val style: InputComponentStyle
+        private val style: CardNumberComponentStyle
     ) : ViewModelProvider.Factory, InjectionClient {
 
         @Inject
