+++ frames/src/main/java/com/checkout/frames/mapper/InputComponentStyleToViewStyleMapper.kt
@@ -3,10 +3,10 @@ package com.checkout.frames.mapper
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.ui.Modifier
 import com.checkout.base.mapper.Mapper
-import com.checkout.frames.style.component.ContainerStyle
-import com.checkout.frames.style.component.InputComponentStyle
-import com.checkout.frames.style.component.InputFieldStyle
-import com.checkout.frames.style.component.TextLabelStyle
+import com.checkout.frames.style.component.base.ContainerStyle
+import com.checkout.frames.style.component.base.InputComponentStyle
+import com.checkout.frames.style.component.base.InputFieldStyle
+import com.checkout.frames.style.component.base.TextLabelStyle
 import com.checkout.frames.style.view.InputComponentViewStyle
 import com.checkout.frames.style.view.InputFieldViewStyle
 import com.checkout.frames.style.view.TextLabelViewStyle
