+++ frames/src/main/java/com/checkout/frames/mapper/TextLabelStyleToStateMapper.kt
@@ -3,8 +3,8 @@ package com.checkout.frames.mapper
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import com.checkout.base.mapper.Mapper
-import com.checkout.frames.style.component.ImageStyle
-import com.checkout.frames.style.component.TextLabelStyle
+import com.checkout.frames.style.component.base.ImageStyle
+import com.checkout.frames.style.component.base.TextLabelStyle
 import com.checkout.frames.view.TextLabelState
 
 internal class TextLabelStyleToStateMapper(
