+++ frames/src/main/java/com/checkout/frames/mapper/TextLabelStyleToViewStyleMapper.kt
@@ -4,8 +4,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.sp
 import com.checkout.base.mapper.Mapper
-import com.checkout.frames.style.component.ContainerStyle
-import com.checkout.frames.style.component.TextLabelStyle
+import com.checkout.frames.style.component.base.ContainerStyle
+import com.checkout.frames.style.component.base.TextLabelStyle
 import com.checkout.frames.style.view.TextLabelViewStyle
 import com.checkout.frames.utils.extensions.toComposeFontWeight
 import com.checkout.frames.utils.extensions.toComposeFontStyle
