+++ frames/src/main/java/com/checkout/frames/style/component/CardNumberComponentStyle.kt
@@ -1,0 +1,7 @@
+package com.checkout.frames.style.component
+
+import com.checkout.frames.style.component.base.InputComponentStyle
+
+public data class CardNumberComponentStyle(
+    val inputStyle: InputComponentStyle
+)
