+++ frames/src/main/java/com/checkout/frames/style/component/base/InputComponentStyle.kt
@@ -1,4 +1,4 @@
-package com.checkout.frames.style.component
+package com.checkout.frames.style.component.base
 
 public data class InputComponentStyle(
     val titleStyle: TextLabelStyle? = null,
