+++ frames/src/main/java/com/checkout/frames/style/component/base/InputFieldStyle.kt
@@ -1,9 +1,9 @@
-package com.checkout.frames.style.component
+package com.checkout.frames.style.component.base
 
 public data class InputFieldStyle(
     val textStyle: TextStyle,
-    val placeholderText: String,
-    val placeholderStyle: TextStyle,
+    val placeholderText: String = "",
+    val placeholderStyle: TextStyle = TextStyle(),
     val containerStyle: ContainerStyle = ContainerStyle(),
     val indicatorStyle: InputFieldIndicatorStyle = InputFieldIndicatorStyle.Border(),
     val leadingIconStyle: ImageStyle? = null,
