+++ frames/src/main/java/com/checkout/frames/style/component/base/TextStyle.kt
@@ -1,4 +1,4 @@
-package com.checkout.frames.style.component
+package com.checkout.frames.style.component.base
 
 import androidx.annotation.ColorLong
 import com.checkout.frames.model.TextAlign
@@ -8,15 +8,15 @@ import com.checkout.frames.model.font.FontWeight
 
 public data class TextStyle(
     /** Text size in sp. */
-    val size: Int,
-    /** The 32-bit ARGB text color. */
+    val size: Int = 14,
+    /** The 32-bit ARGB text color. F.e. black color 0xFF000000. */
     @ColorLong
-    val color: Long,
+    val color: Long = 0xFF000000,
     /** Defines how to align text horizontally. */
     val textAlign: TextAlign = TextAlign.Start,
     /** Defines font family. */
     val font: Font = Font.Default,
-    /** Defines font style, normal or italic. */
+    /** Defines font inputStyle, normal or italic. */
     val fontStyle: FontStyle = FontStyle.Normal,
     /** Defines font weight for the text field. */
     val fontWeight: FontWeight = FontWeight.Normal,
