+++ frames/src/main/java/com/checkout/frames/style/view/TextLabelViewStyle.kt
@@ -44,7 +44,7 @@ import androidx.compose.ui.unit.TextUnit
  * [TextLayoutResult] object that callback provides contains paragraph information, size of the
  * text, baselines and other details. The callback can be used to add additional decoration or
  * functionality to the text. For example, to draw selection around the text.
- * @param style style configuration for the text such as color, font, line height etc.
+ * @param style inputStyle configuration for the text such as color, font, line height etc.
  */
 internal data class TextLabelViewStyle(
     val modifier: Modifier = Modifier,
