+++ frames/src/main/java/com/checkout/frames/view/TextLabel.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import com.checkout.frames.mapper.ImageStyleToComposableImageMapper
 import com.checkout.frames.model.Padding
-import com.checkout.frames.style.component.ImageStyle
+import com.checkout.frames.style.component.base.ImageStyle
 import com.checkout.frames.style.view.TextLabelViewStyle
 
 @Composable
